backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
transformers>=4.45.0,<5.0.0
torch>=2.3.0
accelerate>=0.30.0
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
python-dotenv==1.0.1
rich>=13.0.0
